+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/operations/create_partner_expense_connection.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import connection as shared_connection
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import connection as shared_connection
 from typing import Optional
 
 
@@ -24,8 +23,6 @@ class CreatePartnerExpenseConnectionResponse:
     r"""HTTP response status code for this operation"""
     connection: Optional[shared_connection.Connection] = dataclasses.field(default=None)
     r"""Success"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""