+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/operations/initiate_sync.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import postsync as shared_postsync
+from ...models.shared import postsync as shared_postsync
 from typing import Optional
 
 
@@ -23,8 +22,6 @@ class InitiateSyncResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""