+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.3.0] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [PyPI v0.3.0] https://pypi.org/project/codat-sync-for-expenses-version-1/0.3.0 - previous-versions/sync-for-expenses-version-1
+
+## 2023-11-13 01:24:37
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.4.0] previous-versions/sync-for-expenses-version-1
+### Releases
+- [PyPI v0.4.0] https://pypi.org/project/codat-sync-for-expenses-version-1/0.4.0 - previous-versions/sync-for-expenses-version-1